saving work, stable version

+++ src/app/services/user-service.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { User } from '../user-auth/userregister/user.mode';
+import { stringify } from 'querystring';
 
 @Injectable({
   providedIn: 'root'
@@ -12,17 +13,38 @@ export class UserServiceService {
 
   constructor() { }
 
-  signup(user: User):void{
-    localStorage.setItem( user.email, JSON.stringify(user));
+  signup(user: User): boolean{
+    
     //OLD
     //this.userList.push(user);
-    console.log("este dato está en el navegador",this.datoUsuario);
+    let newUserEmail: string = "";
+    newUserEmail= user.email;
+    if (this.checkIfUserExists(newUserEmail)){
+      return false;
 
+    }else {
+      localStorage.setItem( user.email, JSON.stringify(user));
+      return true;
+    }
+
+    
+    console.log("este dato está en el navegador",this.datoUsuario);
+ 
+
+  }
+
+  checkIfUserExists(userEmail: string): boolean{
+    for(let i=0; i<localStorage.length; i++) {
+      let key = localStorage.key(i);
+      if(key == userEmail){
+        return true;
+      }
+    }
   }
 
   login(user: User):  boolean{
     
-    let incomingUserEmail: string = user.email;
+    //let incomingUserEmail: string = user.email;
     this.datoUsuario = JSON.parse(localStorage.getItem(user.email));
     console.log("entra al login");
     if(this.datoUsuario != null && this.datoUsuario.email == user.email && this.datoUsuario.password == user.password){
